Move window-related functions to GameWindow.hpp
We had a bunch of window-related function (InitD3d*, CreateGameWindow,
WindowProc) in the main.cpp file from early development. However, when
inspecting the mapping, we find that those functions are actually part
of the GameWindow.cpp file.

To be entirely fair, WinMain should _also_ be part of that file... but
we can keep it separate for now.

We also move Clear and SetViewport to ScreenEffect, based on the mapping
again.
