+++ src/GameWindow.hpp
@@ -4,6 +4,9 @@
 #include "inttypes.hpp"
 #include <windows.h>
 
+#define GAME_WINDOW_WIDTH 640
+#define GAME_WINDOW_HEIGHT 480
+
 enum RenderResult
 {
     RENDER_RESULT_KEEP_RUNNING,
@@ -25,6 +28,11 @@ struct GameWindow
     i32 lowPowerActive;
     i32 powerOffActive;
 };
+
+i32 InitD3dInterface();
+void CreateGameWindow(HINSTANCE hInstance);
+i32 InitD3dRendering();
+void InitD3dDevice();
 
 DIFFABLE_EXTERN(GameWindow, g_GameWindow)
 DIFFABLE_EXTERN(i32, g_TickCountToEffectiveFramerate)
