+++ src/ScreenEffect.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <Windows.h>
+#include <d3d8types.h>
+
+void SetViewport(D3DCOLOR color);
